Generative AI Model Files/requirements.txt
transformers>=4.41.0
accelerate>=0.33.0
bitsandbytes>=0.43.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
llama-index>=0.11.0
langchain>=0.2.10
qiskit>=1.0.0
cirq>=1.3.0
pennylane>=0.35.0
numpy>=1.26.0
scipy>=1.11.0
orjson>=3.10.0
tenacity>=8.3.0
fastapi>=0.110.0
uvicorn>=0.29.0
typer>=0.12.0
rich>=13.7.0
pytest>=8.2.0
python-dotenv>=1.0.1
# TODO: Add Python dependencies
# Quantum Computing Framework Dependencies
cirq>=1.2.0
qiskit>=0.45.0
pennylane>=0.32.0

# Backend Framework Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
pydantic>=2.5.0

# Database Dependencies
sqlalchemy>=2.0.0
alembic>=1.13.0

# Utility Dependencies
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.7.0
networkx>=3.2.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.7.0

# Optional: For enhanced quantum simulation
qiskit-aer>=0.13.0
